backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
psycopg2-binary==2.9.10
Flask-Limiter==3.5.0
cryptography==42.0.2
google-cloud-vision==3.7.0
gunicorn==21.2.0
Pillow==10.4.0
reportlab==4.1.0
pytz==2024.1 